Refactor ActionsPanel; fix search
